requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
omdbapi==0.7.0
packaging==23.2
pillow==10.2.0
requests==2.31.0
urllib3==2.2
